bug 1246967 - Add Document.uuid to database
Adds a null uuid column to the Document table. This field is intended to
be used as a strong identifier for an MDN page in the BrowserCompat
project.

This migration needs to be applied in production before the
Django models are updated or the field is used. The database default is
null, so that new documents can be created without error. This will be
fixed in future migrations.